26 organize python scripts

+++ build.sh
@@ -1,10 +1,11 @@
 #!/bin/sh
 
 # Clean gen folder
-rm -r ./resources/generated
+if [ -d './resources/generated']; then
+    rm -r ./resources/generated
+fi
 
 # Run python scripts
 python3 -m pip install -r ./python/requirements.txt
 
-python3 ./python/build_projects.py
-python3 ./python/build_indexes.py
+python3 ./python/build.py
